app/requirements.txt
absl-py==0.12.0
aerich==0.5.3
aiofiles==0.6.0
aiohttp==3.7.4.post0
aioice==0.7.5
aioredis==1.3.1
aiosqlite==0.16.1
alembic==1.5.8
aniso8601==7.0.0
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.22.0
attrs==20.3.0
av==8.0.3
bcrypt==3.2.0
bidict==0.21.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
ciso8601==2.1.3
click==7.1.2
crc32c==2.2
cryptography==3.4.7
databases==0.4.3
ddlparse==1.9.0
dictdiffer==0.8.1
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
fastapi-jwt-auth==0.5.0
FastAPI-SQLAlchemy==0.2.1
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
h11==0.12.0
h5py==2.10.0
hiredis==2.0.0
httptools==0.1.1
idna==2.10
iniconfig==1.1.1
iso8601==0.1.14
Keras-Preprocessing==1.1.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
multidict==5.1.0
netifaces==0.10.9
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.9
passlib==1.7.4
pluggy==0.13.1
promise==2.3
protobuf==3.15.8
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.1
pyee==8.1.0
PyJWT==1.7.1
pylibsrtp==0.6.8
pyparsing==2.4.7
pypika-tortoise==0.1.0
pytest==6.2.3
python-dateutil==2.8.1
python-dotenv==0.17.0
python-editor==1.0.4
python-engineio==4.1.0
python-jose==3.2.0
python-multipart==0.0.5
python-socketio==5.2.1
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
starlette==0.13.6
termcolor==1.1.0
toml==0.10.2
tortoise-orm==0.17.2
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
XlsxWriter==1.3.8
yarl==1.6.3
